+++ app/Http/Controllers/ItemController.php
@@ -11,12 +11,18 @@ class ItemController extends Controller
 {
     public function index()
     {
+        if (!auth()->user()->is_admin){
+            return redirect()->route('dashboard');
+        }
         $items = Item::orderBy('status', 'DESC')->get();
         return view('pages.item.index', compact('items'));
     }
 
     public function showAdd()
     {
+        if (!auth()->user()->is_admin){
+            return redirect()->route('dashboard');
+        }
         $categories = Category::all();
         $suppliers = Supplier::all();
         return view('pages.item.add', compact('categories', 'suppliers'));
@@ -24,6 +30,9 @@ class ItemController extends Controller
 
     public function store(Request $request)
     {
+        if (!auth()->user()->is_admin){
+            return redirect()->route('dashboard');
+        }
         $request->validate([
             'name' => 'required',
             'serial_number' => 'required',
@@ -31,20 +40,27 @@ class ItemController extends Controller
             'category_id' => 'required',
             'supplier_id' => 'required',
         ]);
-
+        //dd($request);
         Item::create([
             'name' => $request->name,
             'serial_number' => $request->serial_number,
             'status' => $request->status,
+            'description' => $request->description,
+            'author' => $request->author,
+            'price' => $request->price,
+            'image' => $request->image,
             'category_id' => $request->category_id,
             'supplier_id' => $request->supplier_id
         ]);
 
-        return redirect()->route('item')->with(['message' => 'Item added', 'alert' => 'alert-success']);
+        return redirect()->route('item')->with(['message' => 'Book added', 'alert' => 'alert-success']);
     }
 
     public function destroy($id)
     {
+        if (!auth()->user()->is_admin){
+            return redirect()->route('dashboard');
+        }
         $item = Item::find($id);
         $item->delete();
 
@@ -53,6 +69,9 @@ class ItemController extends Controller
 
     public function showEdit($id)
     {
+        if (!auth()->user()->is_admin){
+            return redirect()->route('dashboard');
+        }
         $item = Item::find($id);
         $categories = Category::all();
         $suppliers = Supplier::all();
@@ -62,6 +81,9 @@ class ItemController extends Controller
 
     public function update($id, Request $request)
     {
+        if (!auth()->user()->is_admin){
+            return redirect()->route('dashboard');
+        }
         $item = Item::find($id);
 
         $request->validate([
